Refactoring for public packaging and distribution, including Sphinx docs.
