+++ docs/source/index.rst
@@ -1,0 +1,21 @@
+survey-eval Python Package Documentation
+========================================
+
+.. include:: ../../README.rst
+    :start-after: ==========
+     surveyeval
+     ==========
+    :end-before: Full documentation
+
+.. toctree::
+    :maxdepth: 5
+    :caption: Reference
+
+    modules
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
