+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+surveyeval
+==========
+
+.. toctree::
+   :maxdepth: 4
+
+   surveyeval
