+++ docs/source/surveyeval.core_evaluation_lenses.rst
@@ -1,0 +1,6 @@
+surveyeval.core\_evaluation\_lenses module
+==========================================
+
+.. automodule:: surveyeval.core_evaluation_lenses
+   :members:
+   :show-inheritance:
