+++ docs/source/surveyeval.evaluation_engine.rst
@@ -1,0 +1,6 @@
+surveyeval.evaluation\_engine module
+====================================
+
+.. automodule:: surveyeval.evaluation_engine
+   :members:
+   :show-inheritance:
