+++ docs/source/surveyeval.rst
@@ -1,0 +1,19 @@
+surveyeval package
+==================
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   surveyeval.core_evaluation_lenses
+   surveyeval.evaluation_engine
+   surveyeval.survey_parser
+
+Module contents
+---------------
+
+.. automodule:: surveyeval
+   :members:
+   :show-inheritance:
