+++ docs/source/surveyeval.survey_parser.rst
@@ -1,0 +1,6 @@
+surveyeval.survey\_parser module
+================================
+
+.. automodule:: surveyeval.survey_parser
+   :members:
+   :show-inheritance:
